requirements.txt
# Base ----------------------------------------
sqlalchemy==1.4.37
pandas==1.4
python-dotenv==0.20.0
psycopg2-binary==2.9.3

numpy>=1.21.6
# opencv-python>=4.1.2
# matplotlib>=3.2.2
# ipyleaflet>=0.14.0
# sympy>=1.10.1
jupyter==1.*

# Machine learning ----------------------------
# scikit-learn
